Added a message base atom and made the Alert and Message atoms use it.
Refactored the Message atom back into a React component to allow the variant prop to be used in place of the color prop.
