+++ src/components/atoms/_base/index.js
@@ -4,3 +4,4 @@ export { baseControl, control, select } from './control'
 export { flex } from './flex'
 export { grid } from './grid'
 export { list, listTypes, listDefaults, getItemColors } from './list'
+export { message, messageVariants } from './message'
