+++ src/components/atoms/alert/styles.js
@@ -1,31 +1,37 @@
 import { css } from '@emotion/core'
 import styled from '@emotion/styled'
 
-import {
-  bordered,
-  chromatic,
-  margined,
-  padded,
-  typographic,
-} from '../../../traits'
+import { bordered, inline, margined, padded } from '../../../traits'
+
+import { message } from '../_base'
 
 const baseStyles = [
+  ...message.styles(),
   bordered.styles,
-  chromatic.styles,
   margined.styles,
   padded.styles,
-  typographic.styles,
-  () => {
-    return css`
-      display: flex;
-      align-items: center;
-      .icon {
-        margin-right: 1em;
-      }
-    `
-  },
+  () => css`
+    align-items: center;
+    .icon {
+      margin-right: 1em;
+    }
+  `,
 ]
 
-export const AlertBlock = styled.div(...baseStyles)
+export const AlertBlock = styled.div(
+  ...baseStyles,
+  () =>
+    css`
+      display: flex;
+    `
+)
 
-export const AlertInline = styled.span(...baseStyles)
+export const AlertInline = styled.span(
+  ...baseStyles,
+  inline.styles,
+  () =>
+    css`
+      display: inline-flex;
+      vertical-align: middle;
+    `
+)
