+++ src/components/atoms/message/index.jsx
@@ -1,0 +1,19 @@
+import React from 'react'
+
+import { message } from '../_base'
+import * as Styled from './styles'
+
+const Message = ({ variant, ...others }) => (
+  <Styled.Message {...others} color={variant} />
+)
+
+Message.propTypes = {
+  ...message.propTypes(),
+}
+
+Message.defaultProps = {
+  ...message.defaultProps(),
+  font: 'message',
+}
+
+export default Message
