+++ src/components/atoms/message/styles.js
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled'
+
+import { message } from '../_base'
+
+// eslint-disable-next-line import/prefer-default-export
+export const Message = styled.label(...message.styles())
